requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
openai==1.57.4
anthropic>=0.40.0
replicate>=1.0.4
Pillow==11.0.0
requests==2.32.3
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0
apscheduler>=3.10.0,<4.0
pytz>=2024.1
